+++ neurocaps/__init__.py
@@ -3,4 +3,4 @@ from . import analysis, extraction
 __all__=["analysis", "extraction"]
 
 # Version in single place
-__version__ = "0.14.6"
+__version__ = "0.14.7"